[AOTI] Improve the two-pass wrapper codegen
Summary: For the second-pass, we don't have to rerun the whole inductor flow again. This PR moves that second-pass to the codegen time. This change not only speeds up the compilation, but also removes kernel scheduling inconsistency between the two passes. Another future improvement is to make the second-pass reuse the scheduler and do the wrapper codegen only.

ghstack-source-id: 1e4c272f29de6b513f6e800b2763c74ab55e479a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113762